should be uint16

+++ kernel/pic.h
@@ -68,7 +68,7 @@ void pic_mask_all_interrupts(void) {
    outb(PIC_SLAVE_DATA, 0xff);
 }
 
-void pic_send_end_of_interrupt(int command_port) {
+void pic_send_end_of_interrupt(uint16_t command_port) {
    // set bit 5 of OCW 2
    outb(command_port, (1<<5));
 }
